storage/decoder/requirements.txt
pillow==11.0.0
numpy==2.2.0
pytest==8.3.4
torch==2.5.1
torchvision==0.20.1
timm==1.0.12
clip @ git+https://github.com/openai/CLIP.git
ftfy==6.3.1 
regex==2024.9.11
tqdm==4.66.6
pika==1.3.2
psycopg2-binary==2.9.10
